+++ app/src/main/java/com/example/moyomoyo/studyrooms/StudyRoomListAdapter.java
@@ -1,11 +1,9 @@
-package com.example.moyomoyo.studyroom;
+package com.example.moyomoyo.studyrooms;
 
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.BaseAdapter;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
